Improvements after human review
