+++ CLAUDE.md
@@ -7,3 +7,6 @@ This project's agent guidance lives in various `AGENTS.md` files (shared by all 
 In particular, see the **Skills** section: when a request matches a skill, open the relevant
 `.github/skills/<name>/SKILL.md` and follow it. For example, when the user gives a YouTrack
 issue id like `BL-1234`, follow `.github/skills/youtrack-fix/SKILL.md`.
+
+Whenever a task adds, modifies, or reviews localizable strings (XLF entries), automatically
+follow `.github/skills/xlf-strings/SKILL.md` — no explicit request needed.
